Fix #7140 - [Bug] Cannot merge branch when contains smartfolder

+++ becpg-core/src/main/java/fr/becpg/repo/entity/impl/EntityServiceImpl.java
@@ -1,18 +1,18 @@
 /*******************************************************************************
- * Copyright (C) 2010-2018 beCPG. 
- *  
- * This file is part of beCPG 
- *  
- * beCPG is free software: you can redistribute it and/or modify 
- * it under the terms of the GNU Lesser General Public License as published by 
- * the Free Software Foundation, either version 3 of the License, or 
- * (at your option) any later version. 
- *  
- * beCPG is distributed in the hope that it will be useful, 
- * but WITHOUT ANY WARRANTY; without even the implied warranty of 
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the 
- * GNU Lesser General Public License for more details. 
- *  
+ * Copyright (C) 2010-2018 beCPG.
+ *
+ * This file is part of beCPG
+ *
+ * beCPG is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * beCPG is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
  * You should have received a copy of the GNU Lesser General Public License along with beCPG. If not, see <http://www.gnu.org/licenses/>.
  ******************************************************************************/
 package fr.becpg.repo.entity.impl;
@@ -39,6 +39,7 @@ import org.alfresco.model.ContentModel;
 import org.alfresco.repo.coci.CheckOutCheckInServiceImpl;
 import org.alfresco.repo.content.MimetypeMap;
 import org.alfresco.repo.security.authentication.AuthenticationUtil;
+import org.alfresco.repo.virtual.VirtualContentModel;
 import org.alfresco.service.cmr.model.FileFolderService;
 import org.alfresco.service.cmr.model.FileInfo;
 import org.alfresco.service.cmr.repository.ChildAssociationRef;
@@ -61,18 +62,18 @@ import fr.becpg.common.BeCPGException;
 import fr.becpg.model.BeCPGModel;
 import fr.becpg.model.EntityListState;
 import fr.becpg.model.ReportModel;
-import fr.becpg.model.SystemState;
 import fr.becpg.repo.RepoConsts;
 import fr.becpg.repo.entity.EntityDictionaryService;
 import fr.becpg.repo.entity.EntityListDAO;
 import fr.becpg.repo.entity.EntityService;
+import fr.becpg.repo.entity.datalist.impl.SmartContentExtractor;
 import fr.becpg.repo.helper.TranslateHelper;
 
 /**
  * Entity Service implementation
- * 
+ *
  * @author querephi
- * 
+ *
  */
 @Service("entityService")
 public class EntityServiceImpl implements EntityService {
@@ -97,14 +98,12 @@ public class EntityServiceImpl implements EntityService {
 	@Autowired
 	private ContentService contentService;
 
-//	@Autowired
-//	private DictionaryService dictionaryService;
-//	
-	
+	// @Autowired
+	// private DictionaryService dictionaryService;
+	//
+
 	@Autowired
 	private EntityDictionaryService entityDictionaryService;
-	
-	
 
 	private static Integer MAX_DEPTH_LEVEL = 6;
 
@@ -114,10 +113,9 @@ public class EntityServiceImpl implements EntityService {
 		IGNORE_PARENT_ASSOC_TYPES.add(ContentModel.ASSOC_IN_ZONE);
 	}
 
-
 	/**
 	 * Load an image in the folder Images.
-	 * 
+	 *
 	 * @param nodeRef
 	 *            the node ref
 	 * @param imgName
@@ -146,7 +144,7 @@ public class EntityServiceImpl implements EntityService {
 	}
 
 	/**
-	 * 
+	 *
 	 * @param nodeRef
 	 * @param imgName
 	 * @return List of images nodeRefs
@@ -191,11 +189,12 @@ public class EntityServiceImpl implements EntityService {
 
 	/**
 	 * Load the image associated to the node.
-	 * 
+	 *
 	 * @param nodeRef
 	 *            the node ref
 	 * @return the image
 	 */
+	@Override
 	public byte[] getImage(NodeRef nodeRef) {
 
 		byte[] imageBytes = null;
@@ -302,7 +301,7 @@ public class EntityServiceImpl implements EntityService {
 	@Override
 	public String getDefaultImageName(QName entityTypeQName) {
 		String imgName = TranslateHelper.getTranslatedPath(RepoConsts.PATH_LOGO_IMAGE + "." + entityTypeQName.getLocalName());
-		if (imgName == null || imgName.isEmpty()) {
+		if ((imgName == null) || imgName.isEmpty()) {
 			imgName = TranslateHelper.getTranslatedPath(RepoConsts.PATH_LOGO_IMAGE);
 		}
 		if (imgName != null) {
@@ -335,12 +334,12 @@ public class EntityServiceImpl implements EntityService {
 	public NodeRef getOrCreateDocumentsFolder(NodeRef entityNodeRef) {
 		return getDocumentsFolder(entityNodeRef, true);
 	}
-	
+
 	@Override
 	public NodeRef getDocumentsFolder(NodeRef entityNodeRef, boolean create) {
 		String documentsFolderName = TranslateHelper.getTranslatedPath(RepoConsts.PATH_DOCUMENTS);
 		NodeRef documentsFolderNodeRef = nodeService.getChildByName(entityNodeRef, ContentModel.ASSOC_CONTAINS, documentsFolderName);
-		if (documentsFolderNodeRef == null && create) {
+		if ((documentsFolderNodeRef == null) && create) {
 			documentsFolderNodeRef = fileFolderService.create(entityNodeRef, documentsFolderName, ContentModel.TYPE_FOLDER).getNodeRef();
 		}
 		return documentsFolderNodeRef;
@@ -352,17 +351,14 @@ public class EntityServiceImpl implements EntityService {
 		Map<QName, Serializable> props = new HashMap<>();
 		props.put(ContentModel.PROP_NAME, entityName);
 
-		if (sourceNodeRef != null && nodeService.exists(sourceNodeRef)) {
+		if ((sourceNodeRef != null) && nodeService.exists(sourceNodeRef)) {
 			logger.debug("Copy existing entity");
 
-			ret = AuthenticationUtil.runAsSystem(new AuthenticationUtil.RunAsWork<NodeRef>() {
-				@Override
-				public NodeRef doWork() throws Exception {
-					NodeRef ret = copyService.copyAndRename(sourceNodeRef, parentNodeRef, ContentModel.ASSOC_CONTAINS, ContentModel.ASSOC_CHILDREN,
-							true);
-					nodeService.setProperty(ret, ContentModel.PROP_NAME, entityName);
-					return ret;
-				}
+			ret = AuthenticationUtil.runAsSystem(() -> {
+				NodeRef ret1 = copyService.copyAndRename(sourceNodeRef, parentNodeRef, ContentModel.ASSOC_CONTAINS, ContentModel.ASSOC_CHILDREN,
+						true);
+				nodeService.setProperty(ret1, ContentModel.PROP_NAME, entityName);
+				return ret1;
 			});
 
 		} else {
@@ -383,7 +379,7 @@ public class EntityServiceImpl implements EntityService {
 	/**
 	 * Get original name from the working copy name and the cm:workingCopyLabel
 	 * that was used to create it.
-	 * 
+	 *
 	 * @param workingCopyLabel
 	 * @return original name
 	 */
@@ -410,7 +406,7 @@ public class EntityServiceImpl implements EntityService {
 
 	private void copyOrMoveFiles(NodeRef sourceNodeRef, NodeRef targetNodeRef, boolean isCopy) {
 
-		if (targetNodeRef != null && sourceNodeRef != null) {
+		if ((targetNodeRef != null) && (sourceNodeRef != null) && !nodeService.hasAspect(sourceNodeRef, VirtualContentModel.ASPECT_VIRTUAL)) {
 
 			for (FileInfo file : fileFolderService.list(sourceNodeRef)) {
 
@@ -465,22 +461,25 @@ public class EntityServiceImpl implements EntityService {
 	@Override
 	public void deleteFiles(NodeRef entityNodeRef, boolean deleteArchivedNodes) {
 
-		if (entityNodeRef != null) {
+		if (entityNodeRef != null && !nodeService.hasAspect(entityNodeRef, VirtualContentModel.ASPECT_VIRTUAL)) {
 			for (FileInfo file : fileFolderService.list(entityNodeRef)) {
-
-				logger.debug("delete file: " + file.getName() + " entityFolderNodeRef: " + entityNodeRef);
+				if (logger.isDebugEnabled()) {
+					logger.debug("delete file: " + file.getName() + " entityFolderNodeRef: " + entityNodeRef);
+				}
 				deleteNode(file.getNodeRef(), deleteArchivedNodes);
 			}
 		}
 	}
 
 	private void deleteNode(NodeRef nodeRef, boolean deleteArchivedNode) {
-
-		// delete from trash
-		if (deleteArchivedNode) {
-			nodeService.addAspect(nodeRef, ContentModel.ASPECT_TEMPORARY, null);
+		// Test folder exists
+		if (nodeService.exists(nodeRef)) {
+			// delete from trash
+			if (deleteArchivedNode) {
+				nodeService.addAspect(nodeRef, ContentModel.ASPECT_TEMPORARY, null);
+			}
+			nodeService.deleteNode(nodeRef);
 		}
-		nodeService.deleteNode(nodeRef);
 	}
 
 	@Override
@@ -491,14 +490,11 @@ public class EntityServiceImpl implements EntityService {
 			deleteNode(listContainerNodeRef, deleteArchivedNodes);
 		}
 	}
-	
-
 
 	@Override
 	public NodeRef getEntityNodeRef(NodeRef nodeRef, QName itemType) {
 		return getEntityNodeRef(nodeRef, itemType, new HashSet<NodeRef>());
 	}
-	
 
 	private NodeRef getEntityNodeRef(NodeRef nodeRef, QName itemType, Set<NodeRef> visitedNodeRefs) {
 		if (nodeService.exists(nodeRef)) {
@@ -553,8 +549,7 @@ public class EntityServiceImpl implements EntityService {
 		}
 
 		return true;
-		
-	}
 
+	}
 
 }
